\name{ncradPage-package}
\alias{ncradPage-package}
\alias{ncradPage}
\docType{package}
\title{
\packageTitle{ncradPage}
}
\description{
\packageDescription{ncradPage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ncradPage}
\packageIndices{ncradPage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ncradPage}

Maintainer: \packageMaintainer{ncradPage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
